fix for missing int cast which cause compilation warning/error when ctype,__cplusplus and esp32 is used error: array subscript has type 'char' [-Werror=char-subscripts]

+++ 309/CO_gateway_ascii.c
@@ -564,7 +564,7 @@ static inline void convertToLower(char *token, size_t maxCount) {
         if (*c == 0) {
             break;
         } else {
-            *c = tolower(*c);
+            *c = tolower((int)*c);
         }
         c++;
     }
@@ -657,7 +657,7 @@ void CO_GTWA_process(CO_GTWA_t *gtwa,
                 /* empty token, break on error */
                 err = true;
                 break;
-            } else if (isdigit(tok[0]) == 0) {
+            } else if (isdigit((int)tok[0]) == 0) {
                 /* <command> found */
                 break;
             } else if (closed != 0) {
